requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-chroma
pypdf
sentence-transformers
langchain-groq